🎨 main header, layout refactor
- MainHeader 에서 fetch하는 코드들 제거
- 초기 fetch들은 main layout으로 통일
- 각각의 fetch함수들은 파일 분리
