+++ src/app/main/_header.tsx
@@ -1,50 +1,32 @@
 'use client';
 
 import Link from 'next/link';
-import { Dispatch, SetStateAction, useContext, useEffect, useRef, useState } from 'react';
+import { useContext, useEffect, useRef, useState } from 'react';
 import { FaInfoCircle, FaUser } from 'react-icons/fa';
 import DialogModalTwoButton from '@/app/_components/modalTwoButton';
-import DialogModalOneButton from '@/app/_components/modalOneButton';
 import { refreshJwt } from '@/utils/refreshJwt/refresh-jwt-token';
 import { logout } from '@/utils/logout/logout';
-import { MyProfileContext, MyProfileEv } from '@/app/main/_profileContext';
-import { userProfileMeDto } from '@/app/_dto/fetch-profile/Profile.dto';
-import { Logger } from '@/utils/logger/Logger';
+import { MyProfileContext } from '@/app/main/_profileContext';
 import { FaXmark } from 'react-icons/fa6';
 import WebSocketState from '../_components/webSocketState';
 import { NotificationContext } from './layout';
 import { webSocketManager } from '@/app/main/_websocketManager';
-import { fetchMyProfile } from '@/utils/profile/fetchMyProfile';
 
 type headerProps = {
-  setUserProfile: Dispatch<SetStateAction<userProfileMeDto | undefined>>;
+  questionsNum: number;
+  loginChecked: boolean;
 };
-export default function MainHeader({ setUserProfile }: headerProps) {
+export default function MainHeader({ questionsNum, loginChecked }: headerProps) {
   const profile = useContext(MyProfileContext);
   const logoutModalRef = useRef<HTMLDialogElement>(null);
-  const forcedLogoutModalRef = useRef<HTMLDialogElement>(null);
-  const [questionsNum, setQuestions_num] = useState<number>(0);
   const [questionsToastMenu, setQuestionsToastMenu] = useState<boolean>(false);
   const websocketRef = useRef<WebSocket | null>(null);
   const [wsState, setWsState] = useState<number | undefined>();
   const ws_retry_counter = useRef<number>(0);
-  const [loginChecked, setLoginChecked] = useState<boolean>(false);
   const toastTimeout = useRef<NodeJS.Timeout>();
   const [notiNum, setNotiNum] = useState<number>(0);
 
   const notificationContext = useContext(NotificationContext);
-
-  const onProfileUpdateEvent = (ev: CustomEvent<Partial<userProfileMeDto>>) => {
-    const logger = new Logger('onProfileUpdateEvent', { noColor: true });
-    setUserProfile((prev) => {
-      if (prev) {
-        const newData = { ...prev, ...ev.detail };
-        logger.log('My Profile Context Update With: ', ev.detail);
-        return newData;
-      }
-    });
-    setQuestions_num((prev) => ev.detail.questions ?? prev);
-  };
 
   const menuClose = () => {
     const el = document.activeElement as HTMLLIElement;
@@ -97,27 +79,6 @@ export default function MainHeader({ setUserProfile }: headerProps) {
     if (!notificationContext) return;
     setNotiNum(notificationContext.unread_count);
   }, [notificationContext]);
-
-  useEffect(() => {
-    const onResNotOk = () => {
-      forcedLogoutModalRef.current?.showModal();
-    };
-    if (setUserProfile) {
-      fetchMyProfile(onResNotOk).then((r) => {
-        setUserProfile(r);
-        setQuestions_num(r?.questions ?? 0);
-        setLoginChecked(true);
-      });
-    }
-  }, [setUserProfile]);
-
-  useEffect(() => {
-    MyProfileEv.addEventListener(onProfileUpdateEvent);
-
-    return () => {
-      MyProfileEv.removeEventListener(onProfileUpdateEvent);
-    };
-  }, []);
 
   useEffect(() => {
     const fn = async () => {
@@ -224,13 +185,6 @@ export default function MainHeader({ setUserProfile }: headerProps) {
         confirmButtonText={'로그아웃'}
         cancelButtonText={'취소'}
         ref={logoutModalRef}
-        onClick={logout}
-      />
-      <DialogModalOneButton
-        title={'자동 로그아웃'}
-        body={'로그인 유효시간이 만료되어서 로그아웃 되었어요!'}
-        buttonText={'확인'}
-        ref={forcedLogoutModalRef}
         onClick={logout}
       />
       <div
